move ProtocolFeature and checked_feature to near-primitives-core
