+++ core/primitives/src/epoch_manager.rs
@@ -1,5 +1,4 @@
 use crate::challenge::SlashedValidator;
-use crate::checked_feature;
 use crate::num_rational::Rational32;
 use crate::shard_layout::ShardLayout;
 use crate::types::validator_stake::ValidatorStakeV1;
@@ -9,6 +8,7 @@ use crate::types::{
 };
 use crate::version::PROTOCOL_VERSION;
 use borsh::{BorshDeserialize, BorshSerialize};
+use near_primitives_core::checked_feature;
 use near_primitives_core::hash::CryptoHash;
 use near_primitives_core::types::BlockHeight;
 use smart_default::SmartDefault;
@@ -438,8 +438,9 @@ pub mod epoch_info {
     use std::collections::{BTreeMap, HashMap};
 
     use crate::types::validator_stake::ValidatorStakeV1;
-    use crate::{checked_feature, epoch_manager::RngSeed, rand::WeightedIndex};
+    use crate::{epoch_manager::RngSeed, rand::WeightedIndex};
     use near_primitives_core::{
+        checked_feature,
         hash::hash,
         types::{BlockHeight, ShardId},
     };
